fix: hide navbar in home page + resolve build error with vitamin info
