+++ frontend/app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next";
 import { Montserrat_Alternates, Instrument_Sans } from "next/font/google";
 import "./globals.css";
-import { Navbar, MobileNavbar } from "@/components/navbar";
+import { AppShell } from "@/components/navbar";
 
 const montserrat = Montserrat_Alternates({
   subsets: ["latin"],
@@ -43,9 +43,7 @@ export default function RootLayout({
       className={`${montserrat.variable} ${instrument.variable} ${montserratAlternates.variable} ${instrumentSans.variable} h-full antialiased`}
     >
       <body className="min-h-full flex flex-col">
-        <Navbar />
-        <main className="flex-1 pb-20 md:pb-0">{children}</main>
-        <MobileNavbar />
+        <AppShell>{children}</AppShell>
       </body>
     </html>
   );
