+++ frontend/components/navbar/index.ts
@@ -1,2 +1,3 @@
+export { default as AppShell } from "./AppShell";
 export { default as Navbar } from "./Navbar";
 export { default as MobileNavbar } from "./MobileNavbar";
